♻️ Refactor: 랭킹/검색 score 타입 Long → Double 통일
- popularityScore 및 PostRank/CategoryRank/RegionRank score를 Double로 변경
- 검색/랭킹 응답 DTO 및 매퍼 정합성 정리

+++ search-service/src/main/java/com/traveler/search/domain/post/repository/PostDocumentRepositoryCustom.java
@@ -9,8 +9,8 @@ import org.springframework.data.domain.Pageable;
 
 public interface PostDocumentRepositoryCustom {
 
-    record RankingResult(String name, Long score) {
-        public static RankingResult of(String name, Long score) {
+    record RankingResult(String name, Double score) {
+        public static RankingResult of(String name, Double score) {
             return new RankingResult(name, score);
         }
     }
